+++ CHANGELOG.md
@@ -1,9 +1,15 @@
+## 0.3.3
+
+- Add security warning for configuration files
+- Improve (and unit test) debug output
+- Improve test coverage
+
 ## 0.3.2
 
 - Fix severe performance issue, `kubectl` was always called with `--all-namespaces` (#114)
 - Make the troubleshooting guide friendlier + document debug options
-- `--debug cache` now prints name and age of cache filenames
-- `--debug extract` now logs extraction of requests & limits for containers, and capacity for nodes
+- `--debug cache` prints name and age of cache filenames
+- `--debug extract` logs extraction of requests & limits for containers, and capacity for nodes
 - Fix formatting in the PyPI description
 - Add this change log